add report functionality, add some args

+++ check_nc.py
@@ -1,15 +1,16 @@
 from argparse import ArgumentParser
+from os import path as os_path
+
 from git import Repo
 from git import exc
 from git.util import rmtree
-from os import path as os_path
 
-from utils import get_trees
-from utils import get_filenames_with_ext_in_path
-from utils import get_functions_from_trees
-from utils import get_vars_from_trees
-from utils import get_valid_names_from_nodes
-from utils import get_words_from_node_name
+from formatter import format_csv, format_json, format_txt
+from utils import (
+    get_filenames_with_ext_in_path, get_functions_from_trees,
+    get_top_words, get_trees, get_valid_names_from_nodes, get_vars_from_trees,
+    get_words_from_node_name
+)
 
 
 def parse_args():
@@ -35,12 +36,12 @@ def parse_args():
     )
 
     parser.add_argument(
-        '-o', '--output', choices=['stdout', 'json', 'csv'], default='stdout',
-        help='Select destination to output results'
+        '-o', '--format', choices=['txt', 'json', 'csv'], default='txt',
+        help='Select format to output results'
     )
 
     parser.add_argument(
-        '-f', '--file', default='./results.{}',
+        '-f', '--file', default='stdout',
         help='Select file to output results'
     )
 
@@ -51,17 +52,34 @@ def parse_args():
     )
 
     parser.add_argument(
-        '-pl', '--plang', choices=['python'],
+        '-d', '--dir',
+        help='Select dict for analyze - we don\'t clone repo in this case',
+        default='repos',
+    )
+
+    parser.add_argument(
+        '--plang', choices=['python'],
         default='python',
         help='programming language to parse'
     )
 
+    parser.add_argument(
+        '--noclone',
+        help='no clone from repo - take only from dir (-d)',
+        action='store_true'
+    )
+
+    parser.add_argument(
+        '--clear',
+        help='clear repo after script finished',
+        action='store_true'
+    )
     return parser.parse_args()
 
 
-def clone_repository(repo):
+def clone_repository(repo, dest_dir):
     try:
-        repo = Repo.clone_from(repo, os_path.join('.', 'repos'))
+        repo = Repo.clone_from(repo, os_path.join('.', dest_dir))
 
         return repo
     except exc.GitCommandError as e:
@@ -84,16 +102,16 @@ def get_words_in_path(path, tags, node_handler, ext):
     return words
 
 
-def get_top_words(words):
-    return []
+def create_report(words, handler):
+    return handler(words)
 
 
-def create_report(words):
-    return []
-
-
-def output_report(report, output):
-    return
+def output_report(report, filename):
+    if filename == 'stdout':
+        print(report)
+        return
+    with open(filename, 'w') as f:
+        f.write(report)
 
 
 def main():
@@ -108,13 +126,21 @@ def main():
         'func': get_functions_from_trees,
         'var': get_vars_from_trees
     }
+    format_handler = {
+        'txt': format_txt,
+        'json': format_json,
+        'csv': format_csv,
+    }
 
     args = parse_args()
-
-    repo = clone_repository(args.repo)
-    if not repo:
-        exit(1)
-    working_dir = repo.working_dir
+    working_dir = ''
+    if not args.noclone:
+        repo = clone_repository(args.repo, args.dir)
+        if not repo:
+            exit(1)
+        working_dir = repo.working_dir
+    if args.noclone:
+        working_dir = os_path.join('.', args.dir)
 
     if not os_path.exists(working_dir):
         exit(1)
@@ -126,9 +152,10 @@ def main():
         plang[args.plang]
     )
     top_words = get_top_words(words)
-    report = create_report(top_words)
-    output_report(report, args.output)
-    rmtree(working_dir)
+    report = create_report(top_words, format_handler[args.format])
+    output_report(report, args.file.format())
+    if args.clear:
+        rmtree(working_dir)
 
 
 if __name__ == '__main__':
